+++ cmd/zetaclientd/main.go
@@ -82,7 +82,6 @@ func init() {
 	setupGlobalOptions()
 	setupInitializeConfigOptions()
 	setupRelayerOptions()
-	setupInboundOptions()
 
 	// Define commands
 	RootCmd.AddCommand(VersionCmd)